fix: display processing errors and prevent integer overflow
Fix two critical issues:

1. Decode errors silently hidden in status bar: When file decode fails,
   the error message was sent as processing_status, but the status bar
   only displayed it when file_info was present (which it wasn't on error).
   Now display processing_status in the else branch so errors during decode
   are visible to the user with spinner feedback.

2. Integer overflow in progress calculation: For extremely long audio files
   (18+ hours), frames_decoded * 100 would overflow u64. Use u128 for the
   multiplication to safely handle arbitrarily long files without truncation.

Both fixes ensure user visibility of errors and prevent crashes/corruption
on large files.

All 62 tests pass. Zero warnings.

+++ src/audio/decoder.rs
@@ -196,7 +196,8 @@ where
         // Report progress only when total_frames is known and percentage advances
         if let Some(total) = total_frames {
             if total > 0 {
-                let pct = ((frames_decoded * 100) / total).min(100) as u8;
+                // Use u128 to avoid overflow on extremely long audio files (18+ hours).
+                let pct = ((frames_decoded as u128 * 100) / total as u128).min(100) as u8;
                 if pct != last_pct {
                     on_progress(pct);
                     last_pct = pct;
